mueve gradientes a su archivo propio
