+++ lib/features/authentication/presentation/pages/login_page.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../../../core/gradients/gradients.dart';
 import '../../../../core/utils/widgets/generic_scaffold.dart';
-import '../bloc/widgets/login_form.dart';
+import '../widgets/login_page/login_form.dart';
 
 class LoginPage extends StatefulWidget {
   const LoginPage({super.key});
@@ -20,7 +20,7 @@ class _LoginPageState extends State<LoginPage> {
       title: "Login_form_title",
       bodyWidget: Container(
         decoration: BoxDecoration(
-          gradient: Gradients.loginBackGround(),
+          gradient: Gradients.loginGradient(),
         ),
         child: LoginForm(),
       ),
