+++ lib/features/authentication/presentation/pages/register_page.dart
@@ -2,10 +2,10 @@ import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../../../../core/gradients/gradients.dart';
 import '../../../../core/utils/widgets/generic_scaffold.dart';
 import '../bloc/authentication_bloc.dart';
-import '../bloc/widgets/register_background.dart';
-import '../bloc/widgets/register_form.dart';
+import '../widgets/register_page/register_form.dart';
 
 class RegisterPage extends StatefulWidget {
   const RegisterPage({super.key});
@@ -22,15 +22,14 @@ class _RegisterPageState extends State<RegisterPage> {
       showBottomBar: false,
       title: "register_page".tr(),
       bodyWidget: BlocListener<AuthenticationBloc, AuthenticationState>(
-          listener: (context, state) {
-            if (state is RegisterSuccessful) {
-              // Fluttertoast.showToast(
-              //     msg: "REGISTRO EXITOSO!,REDIRIGIENDO", timeInSecForIosWeb: 2);
-            }
-          },
-          child: RegisterBackground(
-            stackBody: RegisterForm(),
-          )),
+        listener: (context, state) {},
+        child: Container(
+          decoration: BoxDecoration(
+            gradient: Gradients.registerGradient(),
+          ),
+          child: RegisterForm(),
+        ),
+      ),
     );
   }
 }
